// packages/logic/src/domain/enums/InheritanceKey.ts
/**
 * 유산 타입, DB에서 사용하는 키
 */
export enum InheritanceKey {
  previous = "previous",
  lived_month = "lived_month",
  max_belong = "max_belong",
  max_domestic_critical = "max_domestic_critical",
  active_action = "active_action",
  //snipe_combat = 'snipe_combat',
  combat = "combat",
  sabotage = "sabotage",
  unifier = "unifier",
  dex = "dex",
  tournament = "tournament",
  betting = "betting",
}